refactor(02-13): hoist FakeHttpMessageHandler to Tests.Shared (IN-03)
- Create `Sources/Tests/ProjectV.Tests.Shared/Helpers/Http/FakeHttpMessageHandler.cs` as the single shared `DelegatingHandler` test seam (public sealed; ctor takes a `Func<HttpRequestMessage, HttpResponseMessage>` responder; exposes a read-only `CallCount`; `SendAsync` invokes the responder, attaches the inbound request to the produced response via `RequestMessage`, and returns `Task.FromResult`).
- Replace the two identical private nested `FakeHttpMessageHandler` declarations in `CommunicationServiceClientTests.cs` and `HttpClientPollyPolicyTests.cs` with `using ProjectV.Tests.Shared.Helpers.Http;` directives. No call-site changes — the public surface (ctor signature, `CallCount` property) matches the original. Preserves the existing comment block about "do not mock HttpMessageHandler with NSubstitute (RESEARCH.md Pitfall 6)" by carrying it into the hoisted file's XML doc.
- Drop redundant `using System.Threading;` from the two callers (now provided by Tests.Shared `SharedUsings.cs` globals).

Core.Tests still 25/25 passing; build + format clean.
